+++ src/main/java/com/example/projectmanagement/modules/designs/app/datastructures/dto/ModuleDefDto.java
@@ -1,0 +1,9 @@
+package com.example.projectmanagement.modules.designs.app.datastructures.dto;
+
+public record ModuleDefDto(
+		Integer id,
+		String moduleName,
+		String programModuleName,
+		String context,
+		Integer projectId) {
+}
