+++ src/main/java/com/example/projectmanagement/modules/designs/app/datastructures/dto/ModuleFeaturesDto.java
@@ -1,0 +1,7 @@
+package com.example.projectmanagement.modules.designs.app.datastructures.dto;
+
+import java.util.List;
+
+public record ModuleFeaturesDto(ModuleDefDto module, List<ApplicationFeatureDto> features) {
+
+}
