+++ src/main/java/com/example/projectmanagement/modules/designs/app/datastructures/form/ModuleFeatureRelationForm.java
@@ -1,0 +1,13 @@
+package com.example.projectmanagement.modules.designs.app.datastructures.form;
+
+import lombok.Data;
+
+@Data
+public class ModuleFeatureRelationForm {
+
+	private Long id;
+	private String description;
+    private Integer featureId;    
+    private Integer moduleId;
+    
+}
