--- src/main/java/com/example/projectmanagement/modules/designs/app/features/datastructures/dto/ApplicationFeatureDto.java
@@ -1,11 +1,0 @@
-package com.example.projectmanagement.modules.designs.app.features.datastructures.dto;
-
-public record ApplicationFeatureDto(
-	    Integer id,
-	    String featureName,
-	    String programFeatureName,
-	    Integer projectId,
-	    Integer historyId,
-	    String description) {
-
-}
