--- src/main/java/com/example/projectmanagement/modules/designs/app/modules/datastructures/dto/ModuleDefDto.java
@@ -1,9 +1,0 @@
-package com.example.projectmanagement.modules.designs.app.modules.datastructures.dto;
-
-public record ModuleDefDto(
-		Integer id,
-		String moduleName,
-		String programModuleName,
-		String context,
-		Integer projectId) {
-}
